Correct anchor ID in ToC link
GitHub automatically generates anchors for every heading in a Markdown file. The hyphen is used as a replacement character in the generated anchor ID.

The "Version History, License" item in the readme table of contents erroneously used an underscore rather than hyphen as the replacement character in the fragment, which caused the link to be non-functional.